requirements.txt
openai>=1.0.0
chromadb>=0.5.0
pypdf>=3.9.0
sentence-transformers>=2.7.0
sounddevice>=0.4.6
numpy>=1.26.0
requests>=2.31.0

# OCR için eklenenler
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.0.0

python-docx>=0.8.11
pandas>=2.2.0
openpyxl>=3.1.2
